Remove filtering on brokers by visibilities
