+++ api/api.go
@@ -126,7 +126,6 @@ func New(ctx context.Context, options *Options) (*web.API, error) {
 			&filters.PatchOnlyLabelsFilter{},
 			filters.NewPlanFilterByVisibility(options.Repository),
 			filters.NewServicesFilterByVisibility(options.Repository),
-			filters.NewBrokersFilterByVisibility(options.Repository),
 		},
 		Registry: health.NewDefaultRegistry(),
 	}, nil
