+++ api/filters/catalog_filter_by_visibility.go
@@ -86,6 +86,40 @@ func (vf *CatalogFilterByVisibility) Run(req *web.Request, next web.Handler) (*w
 	// services := gjson.GetBytes(res.Body, "services")
 }
 
+func servicesCriteria(ctx context.Context, repository storage.Repository, planQuery *query.Criterion) (*query.Criterion, error) {
+	objectList, err := repository.List(ctx, types.ServicePlanType, *planQuery)
+	if err != nil {
+		return nil, err
+	}
+	plans := objectList.(*types.ServicePlans)
+	if plans.Len() < 1 {
+		return nil, nil
+	}
+	serviceIDs := make([]string, 0, plans.Len())
+	for _, p := range plans.ServicePlans {
+		serviceIDs = append(serviceIDs, p.ServiceOfferingID)
+	}
+	c := query.ByField(query.InOperator, "id", serviceIDs...)
+	return &c, nil
+}
+
+func plansCriteria(ctx context.Context, repository storage.Repository, platformID string) (*query.Criterion, error) {
+	objectList, err := repository.List(ctx, types.VisibilityType, query.ByField(query.EqualsOrNilOperator, "platform_id", platformID))
+	if err != nil {
+		return nil, err
+	}
+	visibilityList := objectList.(*types.Visibilities)
+	if visibilityList.Len() < 1 {
+		return nil, nil
+	}
+	planIDs := make([]string, 0, visibilityList.Len())
+	for _, vis := range visibilityList.Visibilities {
+		planIDs = append(planIDs, vis.ServicePlanID)
+	}
+	c := query.ByField(query.InOperator, "id", planIDs...)
+	return &c, nil
+}
+
 // deleteNotVisibleServices deletes the unsupported services from the catalog
 func deleteNotVisibleServices(ctx context.Context, repository storage.Repository, catalog []byte, visiblePlanIDs []string, visibleServiceIDs []string) ([]byte, error) {
 	var err error
